smsc_zabbix_phones/requirements.txt
atlassian-python-api==3.38.0
certifi==2023.5.7
charset-normalizer==3.1.0
Deprecated==1.2.14
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.6
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
urllib3==2.0.3
wrapt==1.15.0
